+++ include/server.h
@@ -13,7 +13,8 @@
 #ifndef FT_SERVER_H
 # define FT_SERVER_H
 
-int		create_server(int port);
-void	stock_in_file(int client_socket);
-void	recv_from_client(int client_socket);
+int			create_server(uint16_t port);
+void		stock_in_file(int client_socket);
+void		recv_from_client(int client_socket);
+
 #endif
